server/requirements.txt
# requirements.txt for Python and GAE

# packages
aiofiles
mypy
email-validator
fastapi
fastapi-utils
google-cloud-storage
gunicorn
#../luncho_python  # don't include for GAE
httpx
json5
mypy-extensions
pycountry
pycountry-convert
pytest
requests
typing-inspect
uvicorn
#orjson does not support Pypy
#ujson does not support Pypy
